Update storage package
