+++ src/main/java/storage/Workout/WorkOutManagerStorage.java
@@ -1,4 +1,4 @@
-package storage;
+package storage.Workout;
 
 import workout.workoutsession.WorkoutSession;
